override access_token changed to 0xFFFFFFFF

+++ indigo_libs/indigo_json.c
@@ -153,7 +153,7 @@ static void *new_text_vector_handler(parser_state state, char *name, char *value
 		} else if (!strcmp(name, "name")) {
 			strncpy(property->name, value, INDIGO_NAME_SIZE);
 		} else if (!strcmp(name, "token")) {
-			property->access_token = (int)strtol(value, NULL, 16);
+			property->access_token = strtol(value, NULL, 16);
 		}
 	} else if (state == END_STRUCT) {
 		indigo_change_property(client, property);
